#805 Store and Retrieve WordEmbeddings on disk cache
- Using float arrays instead of double arrays in Mallet embeddings.

+++ dkpro-core-mallet-asl/src/test/java/de/tudarmstadt/ukp/dkpro/core/mallet/internal/wordembeddings/MalletEmbeddingsUtilsTest.java
@@ -37,7 +37,7 @@ public class MalletEmbeddingsUtilsTest
         int expectedDimensions = 50;
         boolean hasHeader = false;
 
-        Map<String, double[]> embeddings = MalletEmbeddingsUtils
+        Map<String, float[]> embeddings = MalletEmbeddingsUtils
                 .readEmbeddingFileTxt(modelFile, hasHeader);
 
         assertEquals(expectedSize, embeddings.size());
@@ -53,7 +53,7 @@ public class MalletEmbeddingsUtilsTest
         int expectedDimensions = 50;
         boolean hasHeader = true;
 
-        Map<String, double[]> embeddings = MalletEmbeddingsUtils
+        Map<String, float[]> embeddings = MalletEmbeddingsUtils
                 .readEmbeddingFileTxt(modelFile, hasHeader);
 
         assertEquals(expectedSize, embeddings.size());
@@ -69,7 +69,7 @@ public class MalletEmbeddingsUtilsTest
         int expectedDimensions = 50;
         boolean hasHeader = false;
 
-        Map<String, double[]> embeddings = MalletEmbeddingsUtils
+        Map<String, float[]> embeddings = MalletEmbeddingsUtils
                 .readEmbeddingFileTxt(modelFile, hasHeader);
 
         assertEquals(expectedSize, embeddings.size());
